matched func_actor_141000_80133B28 1
State 1 of the main-body table D_actor_141000_80131E58, the step after the
shared turn-to-face body: it rotates the constant local-space offset
D_actor_141000_80131E68 through the root part's matrix into work->step, halving
it first while field_4C8 is clear, then opens the per-axis limit to 0x7FFF and
bumps the state. The shape is the sibling shared body ActorsShared80132920 --
same rotate-then-open-the-limit tail, same u16 counter -- which is what supplied
the field names and the (VECTOR*)&work->step cast.

The seed's whole gap was the m2c collapse of that 16-byte rodata VECTOR into one
scalar load; the copy, not any missing statement, was the 13 instructions. The
0x4C2 counter is u16: the handler reads it lhu while func_actor_141000_80133A00
indexes the table with it lh, so the field's type flips and the dispatcher gains
the (s16) cast the actor_341700/actor_403100 pairs already carry. Rule written
up in DECOMPILATION_LEARNINGS.md.

+++ src/actors/actor_141000/actor_141000_3.c
@@ -48,5 +48,5 @@ void func_actor_141000_80133A00(Task* arg0)
 
     work = (Actor141000Work*)arg0->idMap;
     sp   = D_actor_141000_80131E58;
-    sp.funcs[work->field_4C2](arg0);
+    sp.funcs[(s16)work->field_4C2](arg0);
 }
